The turtle, he shoots grass.

+++ src/com/thecherno/rain/entity/mob/Mob.java
@@ -41,7 +41,7 @@ public abstract class Mob extends Entity {
 	
 	protected void shoot(int x, int y, double dir){
 		//dir = Math.toDegrees(dir);
-		Projectile p = new WizardProjectile(x,y,(int)dir);
+		Projectile p = new WizardProjectile(x,y,dir);
 		projectiles.add(p); 
 		Level.add(p);
